refactor: 将配置管理功能集中到_bootstrap模块
