+++ src/ErisPulse/Core/adapter.py
@@ -144,7 +144,7 @@ class AdapterManager(ManagerBase):
         )
 
         from .router import router
-        from ._self_config import get_server_config
+        from .._bootstrap import get_server_config
         server_config = get_server_config()
 
         host = server_config["host"]
